Update MASTERSOURCE.c
Initially created the base display grid function. A problem we need to tackle (as I have realized) is trying to create a coherent grid whereas even the distance between two nodes are shown. My initial idea is just to output the grid to a text file, but we'll discuss that. Using the same logic that Kyle has created initially, I converted it into a structured type for easier handling.