[SP] Finished Debugging Dijkstra's Algorithm
Issue 1 no longer popped out, which is the rare case that the player gets a shorter distance than Dijkstra's algo.
Issue 2 was given a temporary fix, which is the recursion error that somehow returns an error when not placing a line prior to recursion start.
Issue 2 was fixed by adding the line "int uselessVar = 0;" before the recursion start. It works, but it might be better to convert the recursive function to an iterative function to avoid future bugs and glitches.